Rewrote Code For Efficiency & Created Raindrops Particle System
sorry i wasn't able to do much earlier today! i was doing some school stuff and drew some of the assets. but i basically rewrote the code for the mole game script to be more efficient so we don't have to do the Sprite.Create() stuff or the Resources.Load() stuff, which is a lot of work for a computer to do. I made a Particle System Prefab for the Raindrops which you can check out and use for yourself. It doesn't look the exact same but I tried to make it as close as possible to the original, but you can tweak it if you would like. In addition, I made a TileType script which is just a barebones constructor script for the tiles, where we can add more features to it later. Let me know what you think!