Fixed Sprites PPU, Added Rigidbody2D and Box Colliders for GameObjects. made animation sprite for Mole & Animator Component
The sprites were not at the right ppu before, which is why those little lines were showing up but they're fixed now. The sprites right now are pretty big but we can resize them if needed. Added Rigidbody2D and Box Colliders to the three environment sprites we have along with the mole character since they'll all have some interactability I assume, but I haven't messed with any of the components themselves yet. I also added an Animator component for the Mole GameObject, but I haven't messed with that either yet as I have to finish the rest of the sprites for them plus I have to learn how animation states work in Unity, but it shouldn't be too difficult since there won't be a lot of different states anyways.